Control where `pylint` is imported from. (#145)
* Improve pylint imports.

* Fix tests.
