+++ bundled/tool/runner.py
@@ -13,19 +13,22 @@ import traceback
 # **********************************************************
 # Update sys.path before importing any bundled libraries.
 # **********************************************************
-def update_sys_path(path_to_add: str, append: bool = True) -> None:
+def update_sys_path(path_to_add: str, strategy: str) -> None:
     """Add given path to `sys.path`."""
     if path_to_add not in sys.path and os.path.isdir(path_to_add):
-        if append:
-            sys.path.append(path_to_add)
-        else:
+        if strategy == "useBundled":
             sys.path.insert(0, path_to_add)
+        elif strategy == "fromEnvironment":
+            # no need to append, just pull from the environment.
+            pass
+        else:
+            sys.path.append(path_to_add)
 
 
-# Ensure that we can import LSP libraries, and other bundled linter libraries.
+# Ensure that we can import LSP libraries, and other bundled libraries.
 update_sys_path(
     os.fspath(pathlib.Path(__file__).parent.parent / "libs"),
-    os.getenv("LS_IMPORT_STRATEGY", "fromEnvironment") == "fromEnvironment",
+    os.getenv("LS_IMPORT_STRATEGY", "useBundled"),
 )
